Creating the TestCard app
